app/requirements.txt
fastapi
uvicorn
jinja2
langchain
langchain-openai
langgraph
python-dotenv
pymongo